lxd/storage/drivers/ceph: Rework parseParent
The regexp logic was very hard to follow and also wasn't properly
handling volume names that make use of special characters.

The new logic may be slower but is a lot easier to follow and reason
about as well as being a lot easier to tweak in the future to handle
more edge cases.

Also reworks the test to output something that's easier to check.

Signed-off-by: Stéphane Graber <stgraber@stgraber.org>
Sponsored-by: ActivePort (https://www.activeport.com.au)
(cherry picked from commit 05963ff11587360b12fbabccef9f8bbc43f2f4ff)
Signed-off-by: Simon Deziel <simon.deziel@canonical.com>
License: Apache-2.0